refactored partner service
