+++ src/main/scala/krs/api/ApiModule.scala
@@ -4,4 +4,5 @@ import krs.domain.{ DomainModule }
 
 trait ApiModule { this: DomainModule =>
   val userApi: UserApi = new UserApi(userRepository)
+  val partnerApi: PartnerApi = new PartnerApi(partnerRepository)
 }
