+++ src/main/scala/krs/api/PartnerApi.scala
@@ -1,0 +1,9 @@
+package krs.api
+
+import krs.domain.{ PartnerRepository }
+
+class PartnerApi(partnerRepository: PartnerRepository) {
+  def getOffers() = {
+    partnerRepository.loadOffers()
+  }
+}
