+++ src/main/scala/krs/service/PartnerServer.scala
@@ -5,31 +5,14 @@ import com.twitter.finagle.Thrift
 import com.twitter.finagle.stats.Counter
 import com.twitter.server.TwitterServer
 
-import krs.thriftscala.{ PartnerService, PartnerOffer, OfferResponse }
-import krs.infrastructure.PartnerRepositoryFS
+import krs.service.impl.{ PartnerServerImpl }
 
 object PartnerServer extends TwitterServer {
-  val repo = new PartnerRepositoryFS("./fixtures/offers.json")
-  val offers = repo.loadOffers()
-
-  def buildServer(): PartnerService[Future] = {
-    new PartnerService[Future] {
-      def getOffers() = {
-        val partnerOffers = offers.map((offer) => {
-          PartnerOffer(
-            offer.provider,
-            Option(offer.creditScoreRange.min),
-            Option(offer.creditScoreRange.max))
-        })
-        Future.value(OfferResponse(partnerOffers))
-      }
-    }
-  }
-
+  val serviceImpl = new PartnerServerImpl()
   val partnerService: Counter = statsReceiver.counter("partnerService")
 
   def main(): Unit = {
-    val service = buildServer()
+    val service = serviceImpl()
 
     val server = Thrift.server
       .withStatsReceiver(statsReceiver)
