+++ src/main/scala/krs/service/UserServer.scala
@@ -4,6 +4,7 @@ import com.twitter.util.{ Await }
 import com.twitter.finagle.Thrift
 import com.twitter.finagle.stats.Counter
 import com.twitter.server.TwitterServer
+
 import krs.service.impl.{ UserServerImpl }
 
 object UserServer extends TwitterServer {
